Initial implementation of python test refactoring
