+++ test_collections/sdk_tests/__init__.py
@@ -13,5 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from .support.python_testing import sdk_python_collection
-from .support.yaml_tests import custom_collection, sdk_collection
+
+from ..sdk_tests.support import custom_collection, sdk_collection, sdk_python_collection
