+++ test_collections/sdk_tests/support/models/matter_test_declarations.py
@@ -22,9 +22,51 @@ from app.test_engine.models.test_declarations import (
 )
 from test_collections.sdk_tests.support.models.sdk_test_folder import SDKTestFolder
 
-from .python_test_models import MatterTestType, PythonTest
-from .test_case import PythonTestCase
-from .test_suite import PythonTestSuite, SuiteType
+from ..python_testing.models.python_test_models import PythonTest
+from ..python_testing.models.test_case import PythonTestCase
+from ..python_testing.models.test_suite import PythonTestSuite
+from ..python_testing.models.test_suite import SuiteType as PythonSuiteType
+from ..yaml_tests.models.test_case import YamlTestCase
+from ..yaml_tests.models.test_suite import SuiteType as YamlSuiteType
+from ..yaml_tests.models.test_suite import YamlTestSuite
+from ..yaml_tests.models.yaml_test_models import YamlTest
+from .matter_test_models import MatterTestType
+
+
+class YamlCollectionDeclaration(TestCollectionDeclaration):
+    def __init__(self, folder: SDKTestFolder, name: str) -> None:
+        super().__init__(path=str(folder.path), name=name)
+        self.yaml_version = folder.version
+
+
+class YamlSuiteDeclaration(TestSuiteDeclaration):
+    """Direct initialization for YAML Test Suite."""
+
+    class_ref: Type[YamlTestSuite]
+
+    def __init__(self, name: str, suite_type: YamlSuiteType, version: str) -> None:
+        super().__init__(
+            YamlTestSuite.class_factory(
+                name=name,
+                suite_type=suite_type,
+                yaml_version=version,
+            )
+        )
+
+
+class YamlCaseDeclaration(TestCaseDeclaration):
+    """Direct initialization for YAML Test Case."""
+
+    class_ref: Type[YamlTestCase]
+
+    def __init__(self, test: YamlTest, yaml_version: str) -> None:
+        super().__init__(
+            YamlTestCase.class_factory(test=test, yaml_version=yaml_version)
+        )
+
+    @property
+    def test_type(self) -> MatterTestType:
+        return self.class_ref.yaml_test.type
 
 
 class PythonCollectionDeclaration(TestCollectionDeclaration):
@@ -38,7 +80,7 @@ class PythonSuiteDeclaration(TestSuiteDeclaration):
 
     class_ref: Type[PythonTestSuite]
 
-    def __init__(self, name: str, suite_type: SuiteType, version: str) -> None:
+    def __init__(self, name: str, suite_type: PythonSuiteType, version: str) -> None:
         super().__init__(
             PythonTestSuite.class_factory(
                 name=name,
