+++ test_collections/sdk_tests/support/tests/support/models/test_python_test_declarations.py
@@ -13,45 +13,49 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+# flake8: noqa
+# Ignore flake8 check for this file
 from unittest import mock
 
-from test_collections.sdk_tests.support.yaml_tests.models.test_declarations import (
-    YamlCaseDeclaration,
-    YamlSuiteDeclaration,
+from test_collections.sdk_tests.support.models.matter_test_declarations import (
+    PythonCaseDeclaration,
+    PythonSuiteDeclaration,
 )
-from test_collections.sdk_tests.support.yaml_tests.models.test_suite import SuiteType
-from test_collections.sdk_tests.support.yaml_tests.models.yaml_test_models import (
-    YamlTest,
+from test_collections.sdk_tests.support.python_testing.models.python_test_models import (
+    PythonTest,
+)
+from test_collections.sdk_tests.support.python_testing.models.test_suite import (
+    SuiteType,
 )
 
 
-def test_yaml_suite_declaration() -> None:
+def test_python_suite_declaration() -> None:
     name = "TestName"
     type = SuiteType.AUTOMATED
     version = "SomeVersionStr"
 
     with mock.patch(
-        "test_collections.sdk_tests.support.yaml_tests.models."
-        "test_suite.YamlTestSuite.class_factory"
+        "test_collections.sdk_tests.support.python_testing.models.test_suite.PythonTestSuite.class_factory"
     ) as class_factory, mock.patch(
         "app.test_engine.models.test_declarations.TestSuiteDeclaration.__init__"
     ) as declaration_init:
-        YamlSuiteDeclaration(name=name, suite_type=type, version=version)
+        PythonSuiteDeclaration(name=name, suite_type=type, version=version)
         class_factory.assert_called_once_with(
-            name=name, suite_type=type, yaml_version=version
+            name=name, suite_type=type, python_test_version=version
         )
         declaration_init.assert_called_once()
 
 
-def test_yaml_case_declaration() -> None:
-    test = YamlTest(name="TestTest", config={}, tests=[])
+def test_python_case_declaration() -> None:
+    test = PythonTest(
+        name="TestTest", description="TestTest description", config={}, steps=[]
+    )
     version = "SomeVersionStr"
     with mock.patch(
-        "test_collections.sdk_tests.support.yaml_tests.models."
-        "test_case.YamlTestCase.class_factory"
+        "test_collections.sdk_tests.support.python_testing.models.test_case.PythonTestCase.class_factory"
     ) as class_factory, mock.patch(
         "app.test_engine.models.test_declarations.TestCaseDeclaration.__init__"
     ) as declaration_init:
-        YamlCaseDeclaration(test=test, yaml_version=version)
-        class_factory.assert_called_once_with(test=test, yaml_version=version)
+        PythonCaseDeclaration(test=test, python_test_version=version)
+        class_factory.assert_called_once_with(test=test, python_test_version=version)
         declaration_init.assert_called_once()
